api/requirements.txt
APIFlask==2.1.1
apispec==6.6.1
attrs==23.2.0
black==24.4.2
blinker==1.8.1
click==8.1.7
colorama==0.4.6
Faker==24.14.1
Flask==3.0.3
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.1
mistune==3.0.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
pytest==8.2.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
six==1.16.0
webargs==8.4.0
Werkzeug==3.0.2
